docs<README.md>: 구현 기능 목록 수정, 예외처리 사항 추가
