+++ src/main/java/baseball/Application.java
@@ -4,6 +4,6 @@ public class Application {
     public static void main(String[] args) {
         do {
             Game.play(RandomNumberGenerator.generateThreeRandomDigits());
-        } while (Game.isContinue());
+        } while (Player.isContinue());
     }
 }
