+++ src/main/java/baseball/Game.java
@@ -1,7 +1,5 @@
 package baseball;
 
-import camp.nextstep.edu.missionutils.Console;
-
 public class Game {
 
     public static void play(int[] generatedNumbers) {
@@ -12,7 +10,7 @@ public class Game {
     }
 
     private static String startTurn(int[] generatedNumbers){
-        String guessResult = Player.guessNumber(generatedNumbers, Player.getPlayerInput());
+        String guessResult = Player.guessNumber(generatedNumbers);
         System.out.println(guessResult);
         return guessResult;
     }
@@ -23,13 +21,5 @@ public class Game {
             return true;
         }
         return false;
-    }
-
-    public static boolean isContinue() {
-        System.out.println(SystemMessage.GAME_END_MESSAGE);
-        System.out.println(SystemMessage.RESTART_SELECT_MESSAGE);
-        String input = Console.readLine();
-        if (Validator.isValidateChoice(input)) return input.equals(Constant.CHOICE_ONE);
-        else throw new IllegalArgumentException(SystemMessage.WRONG_INPUT_MESSAGE);
     }
 }
